+++ packages/components-react/projects/components-wrapper/CHANGELOG.md
@@ -7,12 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- `Content Wrapper`
 - Description property to `p-text-field-wrapper`, `p-textarea-wrapper` and `p-select-wrapper`
 - `Link Social`
 
 ### Changed
 - Improve accessibility of error and success states of form elements
 - Aria-invalid attribute of form elements if they are in error state is now managed by component 
+
+### Removed
+- `safe-zone` property of `p-grid` (`Content Wrapper` should be used instead)
 
 ## [1.4.0-rc.1] - 2020-04-27